collect-data/pip-requirements.txt
pylint
prometheus-api-client
pyyaml
matplotlib
kubernetes